re alocation the package of constans and exception into util
