+++ src/main/java/com/consola/lis/controller/LoginRequestController.java
@@ -1,6 +1,6 @@
 package com.consola.lis.controller;
 
-import com.consola.lis.constans.EndpointConstant;
+import com.consola.lis.util.constans.EndpointConstant;
 import com.consola.lis.dto.AuthResponseDTO;
 import com.consola.lis.dto.LoginRequestDTO;
 import com.consola.lis.service.LoginService;
