+++ src/main/java/com/consola/lis/util/exception/JwtExpiredException.java
@@ -1,4 +1,4 @@
-package com.consola.lis.exception;
+package com.consola.lis.util.exception;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
@@ -8,15 +8,13 @@ import org.springframework.http.HttpStatus;
 @EqualsAndHashCode(callSuper = true)
 @Data
 @RequiredArgsConstructor
-public class NotExistingException extends RuntimeException {
-
+public class JwtExpiredException extends RuntimeException {
     private final String code;
     private final HttpStatus status;
 
-    public NotExistingException(String code, HttpStatus status, String message){
+    public JwtExpiredException(String code, HttpStatus status, String message){
         super(message);
         this.status = status;
         this.code = code;
     }
-
 }
