+++ src/main/java/com/consola/lis/util/exception/NotExistingException.java
@@ -1,4 +1,4 @@
-package com.consola.lis.exception;
+package com.consola.lis.util.exception;
 
 import lombok.Data;
 import lombok.EqualsAndHashCode;
@@ -8,13 +8,15 @@ import org.springframework.http.HttpStatus;
 @EqualsAndHashCode(callSuper = true)
 @Data
 @RequiredArgsConstructor
-public class AlreadyExistsException extends RuntimeException {
+public class NotExistingException extends RuntimeException {
+
     private final String code;
     private final HttpStatus status;
 
-    public AlreadyExistsException(String code, HttpStatus status, String message){
+    public NotExistingException(String code, HttpStatus status, String message){
         super(message);
         this.status = status;
         this.code = code;
     }
+
 }
